feat: modal 과 다음 버튼 연결
